requirements.txt
# Core Agent Framework
langgraph>=0.0.40
langchain-core>=0.1.23
langchain-anthropic>=0.1.4
anthropic>=0.18.1

# MCP Integration for Hummingbot
mcp>=1.0.0

# Data Validation & Configuration
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Trading Platform (via HTTP API)
# Hummingbot runs as separate service - interact via HTTP API
ccxt>=4.0.0

# Data & State Management
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
redis>=5.0.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.3
python-dateutil>=2.8.2

# Async & Concurrency
aiohttp>=3.9.1
aiocache>=0.12.0
asyncio>=3.4.3

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# HTTP & API
requests>=2.31.0
httpx>=0.25.0
websocket-client>=1.7.0

# Monitoring & Logging
structlog>=23.2.0
colorlog>=6.8.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0
hypothesis>=6.92.0
faker>=22.0.0

# Code Quality
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
pylint>=3.0.0

# Utilities
uuid>=1.30
pytz>=2023.3
